Header File Updated

+++ sort.h
@@ -21,7 +21,7 @@ typedef struct listint_s
 
 void print_list(const listint_t *list);
 void print_array(const int *array, size_t size);
-void bubble_sort(int *array, size_t size);
+void bubble_sort(int *array, unsigned int size);
 void insertion_sort_list(listint_t **list);
 void swap(int *x, int *y);
 void node_swap(listint_t *nx, listint_t *ny);
